requirements.txt
langchain-openai==0.2.6
python-dotenv==1.0.1
fastapi==0.115.7
langchain==0.3.8
langchain_community==0.3.8
pydantic==2.10.6
uvicorn==0.30.6
langgraph==0.4.8
playwright==1.52.0
ipykernel
beautifulsoup4